RoleEditMappedUser file name refactor
