+++ public/apps/configuration/panels/role-mapping/role-edit-mapped-user.test.tsx
@@ -19,7 +19,7 @@ import { fetchUserNameList } from '../../utils/internal-user-list-utils';
 import { getRoleMappingData, updateRoleMapping } from '../../utils/role-mapping-utils';
 import { ExternalIdentitiesPanel } from './external-identities-panel';
 import { InternalUsersPanel } from './internal-users-panel';
-import { RoleEditMappedUser } from './RoleEditMappedUser';
+import { RoleEditMappedUser } from './role-edit-mapped-user';
 import { RoleMappingDetail } from '../../types';
 
 jest.mock('../../utils/role-mapping-utils');
